requirements.txt
# --- Deep Learning Framework ---
# Note: It's recommended to install PyTorch according to your specific CUDA version from https://pytorch.org/
torch>=2.0.0
torchvision>=0.15.0

# --- Core Dependencies ---
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0

# --- Machine Learning & Vision ---
scikit-learn>=1.3.0
scikit-image>=0.20.0
einops>=0.6.0
timm>=1.0.0

# --- Deep Learning Extensions ---
tqdm>=4.65.0

# --- Third-party GitHub Repositories ---
# OpenAI CLIP
git+https://github.com/openai/CLIP.git
# PyTorch Wavelets
git+https://github.com/fbcotter/pytorch_wavelets.git